move all DTOs in its own package following
