+++ src/main/java/io/zeebe/monitor/rest/dto/FileDto.java
@@ -25,7 +25,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.zeebe.monitor.rest;
+package io.zeebe.monitor.rest.dto;
 
 public class FileDto
 {
