+++ src/main/java/io/zeebe/monitor/rest/dto/MessageDto.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.zeebe.monitor.rest;
+package io.zeebe.monitor.rest.dto;
 
 public class MessageDto {
