+++ src/main/java/io/zeebe/monitor/rest/dto/WorkerDto.java
@@ -1,4 +1,4 @@
-package io.zeebe.monitor.rest;
+package io.zeebe.monitor.rest.dto;
 
 public class WorkerDto {
